Switch to `esp-hal`
This commit removes the esp-idf-hal dependency for the more light-weight
no-std esp-hal crate.
The `main.rs` file, the `.cargo/config.toml` and the `Cargo.toml` file
have been modified in accordance with the esp-template.

Link: https://github.com/esp-rs/esp-template

The code has seen some drastic changes, most importantly in the
handling of GPIOs:
the problem was that there is no way to get a type-erased GPIO that
could switch between input and output modes, and in general there is
less (or at least less straightforward) support for generic programming
over GPIOs.
The solution was to make the GPIOs initialization outside the components
(e.g. before `lcd::driver::bus::Bus` intialized the GPIOs automatically)
and add replace those with traits instead, moving the implementation
outside the module.

This can be seen with the new `lcd::driver::bus::Pins` and
`lcd::driver::Pins` traits, which are now implemented in `main.rs`,
and in the `bus` module, which makes extensive use of (probably too
'clever') macros to reduce the amount of boilerplate code needed.

Another important change was the removal of every error type:
with the esp-hal the GPIO operations never return any error.

The last problem encountered was with the time measuring:
I couldn't find a simple way to somewhat replicate the behaviour of
`std::time::Instant`