Implement LCD driver
This commits adds a new lcd module that implements a Driver for the
the ST7066U.
All the commands in the specs are added in the `lcd::cmd` module.
The `lcd::bus` submodule handles the data bus pins, by chaning their
configuration between input and output when its needed.

The main.rs now initialzies the display and writes "Hello!"