Implement pixel-precise character renderer
This commit restructures the project by:
- moving the driver from `lcd` to `lcd::driver`
- adding three new modules:
  - bitmap: for handling of the pixel bitmaps to store in the CGROM
  - canvas: for operating with pixels instead of characters
  - ddrom: for search in (a copy of) the DDROM data

The `lcd::bitmap::Bitmap` internally represents data as the CGRAM does,
for easier transfer.

The `lcd::ddrom` module contains a copy of the LCD DDROM which allows
searching `Bitmap`s for equality. The DDROM copy has been generated from
the image found in the specs, the program that generated it can be found
here: TODO

The `lcd::canvas::Canvas` can be considered as an array of `Bitmap`s
on which handles characters as the `Driver` does, but, being
pixel-precise, allows for doing operations such as the pixel-by-pixel
shift of the contents.

This new machansims is used in the `main` function to make the `"Hello
World!"` text rotate on the screen. To speed up transfering the data
to the LCD, in-memory copies of the DDRAM and CGRAM are used to update
only the changed values.