Add `timer` module + time measurement + report
This is the first try at having a functioning time measuring system.

The new timer module uses an hardware timer to count the amount of time
passed. This functionality is the used in the main function to count
the amount of time passed between each step in the loop, writing a
continuous report like it did before (the formatting was improved).

                        --- NOTE ---
This doesn't seem like the best approach yet, the timer could be
auto-reloaded (couldn't find a way) and the `TIME_COUNTER` could be
the HW timer itself (by getting access to its internal counter).